security: fix information exposure and refactor HMAC authentication
- Fix critical security issue: Remove detailed exception messages from API responses
- Extract HMAC authentication logic to separate auth.py module
- Add common request parameter validation function
- Improve code organization and reduce duplication
- Maintain comprehensive logging while securing external responses

Addresses security scanner alerts in GitHub PR review feedback
