+++ mock-student-api/main.py
@@ -12,14 +12,12 @@ student information system API endpoints with in-memory data synchronized with i
 from fastapi import FastAPI, HTTPException, Header, Request
 from fastapi.responses import JSONResponse
 from fastapi.middleware.cors import CORSMiddleware
-import hashlib
-import hmac
 import json
 import os
-from datetime import datetime, timedelta, timezone
 from typing import Optional, List, Dict, Any
 from pydantic import BaseModel, Field
 import logging
+from auth import verify_hmac_signature, validate_request_params, STRICT_TIME_CHECK, STRICT_ENCODE_CHECK, MOCK_HMAC_KEY_HEX
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
@@ -40,11 +38,7 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-# Configuration
-MOCK_HMAC_KEY_HEX = os.getenv("MOCK_HMAC_KEY_HEX", "4d6f636b4b657946726f6d48657841424344454647484a4b4c4d4e4f505152535455565758595a")
-STRICT_TIME_CHECK = os.getenv("STRICT_TIME_CHECK", "true").lower() == "true"
-STRICT_ENCODE_CHECK = os.getenv("STRICT_ENCODE_CHECK", "false").lower() == "true"
-TIME_TOLERANCE_MINUTES = int(os.getenv("TIME_TOLERANCE_MINUTES", "5"))
+# Configuration (HMAC settings moved to auth.py)
 
 # Request/Response Models
 class StudentBasicRequest(BaseModel):
@@ -358,76 +352,7 @@ SAMPLE_TERMS = {
     ]
 }
 
-def verify_hmac_signature(
-    authorization: str,
-    request_body: str,
-    content_type: str,
-    encode_type: Optional[str] = None
-) -> bool:
-    """
-    Verify HMAC-SHA256 signature according to university API specification
-    
-    Authorization format: HMAC-SHA256:<TIME>:<ACCOUNT>:<SIGNATURE_HEX>
-    Message format: <TIME> + <REQUEST_JSON> (no whitespace)
-    
-    Note: TIME should be in UTC format (YYYYMMDDHHMMSS)
-    """
-    try:
-        if not authorization.startswith("HMAC-SHA256:"):
-            logger.warning(f"Invalid authorization format: {authorization}")
-            return False
-            
-        parts = authorization[12:].split(":")
-        if len(parts) != 3:
-            logger.warning(f"Invalid authorization parts count: {len(parts)}")
-            return False
-            
-        time_str, account, signature_hex = parts
-        
-        if len(time_str) != 14 or not time_str.isdigit():
-            logger.warning(f"Invalid time format: {time_str}")
-            return False
-            
-        if STRICT_TIME_CHECK:
-            try:
-                request_time = datetime.strptime(time_str, "%Y%m%d%H%M%S").replace(tzinfo=timezone.utc)
-                current_time = datetime.now(timezone.utc)
-                time_diff = abs((current_time - request_time).total_seconds() / 60)
-                
-                if time_diff > TIME_TOLERANCE_MINUTES:
-                    logger.warning(f"Time difference too large: {time_diff} minutes")
-                    return False
-            except ValueError:
-                logger.warning(f"Invalid time format for parsing: {time_str}")
-                return False
-        
-        if STRICT_ENCODE_CHECK and encode_type != "UTF-8":
-            logger.warning(f"Invalid encode type: {encode_type}")
-            return False
-            
-        if content_type != "application/json;charset=UTF-8":
-            logger.warning(f"Invalid content type: {content_type}")
-            return False
-            
-        message = time_str + request_body
-        hmac_key = bytes.fromhex(MOCK_HMAC_KEY_HEX)
-        
-        expected_signature = hmac.new(
-            hmac_key,
-            message.encode('utf-8'),
-            hashlib.sha256
-        ).hexdigest().lower()
-        
-        if signature_hex.lower() != expected_signature:
-            logger.warning(f"Signature mismatch. Expected: {expected_signature}, Got: {signature_hex}")
-            return False
-            
-        logger.info(f"HMAC verification successful for account: {account}")
-        return True
-        
-    except Exception as e:
-        logger.error(f"HMAC verification error: {str(e)}")
-        return False
+# HMAC verification function moved to auth.py
 
 @app.get("/")
 async def root():
@@ -490,17 +415,10 @@ async def get_student_basic_info(
                 }
             )
         
-        if request.account != "scholarship":
-            return JSONResponse(
-                status_code=400,
-                content={"code": 400, "msg": "Invalid account", "data": []}
-            )
-            
-        if request.action != "qrySoaaScholarshipStudent":
-            return JSONResponse(
-                status_code=400,
-                content={"code": 400, "msg": "Invalid action", "data": []}
-            )
+        # Validate request parameters
+        validation_error = validate_request_params(request.account, request.action, "qrySoaaScholarshipStudent")
+        if validation_error:
+            return JSONResponse(status_code=400, content=validation_error)
         
         # Get student data from in-memory store
         student_data = SAMPLE_STUDENTS.get(request.stdcode)
@@ -516,7 +434,7 @@ async def get_student_basic_info(
         logger.error(f"Error in get_student_basic_info: {str(e)}")
         return JSONResponse(
             status_code=500,
-            content={"code": 500, "msg": f"Internal server error: {str(e)}", "data": []}
+            content={"code": 500, "msg": "Internal server error", "data": []}
         )
 
 @app.post("/getsoaascholarshipstudentterm")
@@ -542,17 +460,10 @@ async def get_student_term_info(
                 content={"code": 401, "msg": "HMAC signature verification failed", "data": []}
             )
         
-        if request.account != "scholarship":
-            return JSONResponse(
-                status_code=400, 
-                content={"code": 400, "msg": "Invalid account", "data": []}
-            )
-            
-        if request.action != "qrySoaaScholarshipStudentTerm":
-            return JSONResponse(
-                status_code=400,
-                content={"code": 400, "msg": "Invalid action", "data": []}
-            )
+        # Validate request parameters
+        validation_error = validate_request_params(request.account, request.action, "qrySoaaScholarshipStudentTerm")
+        if validation_error:
+            return JSONResponse(status_code=400, content=validation_error)
         
         # Get student term data from in-memory store
         student_terms = SAMPLE_TERMS.get(request.stdcode, [])
@@ -576,7 +487,7 @@ async def get_student_term_info(
         logger.error(f"Error in get_student_term_info: {str(e)}")
         return JSONResponse(
             status_code=500,
-            content={"code": 500, "msg": f"Internal server error: {str(e)}", "data": []}
+            content={"code": 500, "msg": "Internal server error", "data": []}
         )
 
 if __name__ == "__main__":
